requirements.txt
# === Core Framework ===
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# === HTTP & Scraping ===
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.27.2
newspaper3k==0.2.8
lxml-html-clean==0.4.4

# === AI / ML — Embeddings ===
sentence-transformers==3.1.1
numpy==1.26.4

# === Vector Database ===
faiss-cpu==1.8.0.post1

# === LLM (future STEP 3) ===
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
openai==1.51.0

# === News API ===
newsapi-python==0.2.7

# === Image Processing & OCR (STEP 6) ===
pytesseract
Pillow
python-multipart

# === Database (STEP 7) ===
sqlalchemy

# === Image Processing & OCR (STEP 6) ===
pytesseract
Pillow
python-multipart
